Added a "gpio_setup" component to the omar project so I can start blinking leds
The goal was to add a command to the "console" that would allow me to
toggle the state of the led to be red, green, or blue. This would involve
adding a command, etc.

But before that, I wanted to figure out how to create a new directory and
add sources there... Basically, I wanted to avoid the problem of being
forced to throw everything into the same big bag-of-code directory.

Luckily, I found some documentation here describing how you can structure
esp32 projects:

  https://dl.espressif.com/doc/esp-idf/latest/api-guides/build-system.html

This was helpful. Basically, you have to push all the 'other' source code
directories down into the "components" directory, and populate each one
with an empty "component.mk" file (so the build system can "see" the
component for what it is), and add an "include" directory for header
files. Once you do that, the build system figures out how to modify the
INCLUDEDIRS and so on accordingly.