Grabbed this 'button' component from the ESP32 training git repo
I'm going to use it to implement a button on my HW_ESP32_PICOKIT
devboard. Might be overkill, but it's convenient and is an example of a
more typical esp-idf "component" than my own 'hw_setup' component is (I'm
using the component construct simply as a way to split the omar codebase
into different directories, basically).