Modified iot_button_create_omar() to only call gpio_install_isr_service() once
Figured out over the weekend that all gpio interrupts are tied to the same
esp32 interrupt "service", and so I added a static boolean that only calls
gpio_install_isr_service() once. Subsequent calls to
iot_button_create_omar() simply add the handler for that particular gpio to
the table of callbacks.