Added an "als" menu command to display ambient light sensor reading
This just dumps the raw reading. Seems to respond as expected to covering
it up (goes to 0) or turning on the adjacent LED (shoots up to close to a
full-scale reading of 4096).

It seems pretty sensitive, so I configure the attenuation for the ESP32 ADC
to ADC_ATTEN_DB_11, which at least makes it difficult to overflow the
output.