Verified that I get a DNS server IP address, but still get a null pointer when looking up our Jenkins server
Added a big utility routine called "wifi_event()" that identified each and
every wifi event we get -- just to rule out the possibility that we're not
in the state we think we are.

But our wifi state looks good, and we still can't get a decent IP address
for Jenkins.. Will continue debugging though.