DNS lookup succeeds when I connect to "Currant", but still fails on "Stringer 2.4"
The code has been re-tweaked to retry forever in the event of a failed DNS
query. But no matter how long I retry, "Stringer 2.4" won't give me the
Jenkins echo server IP address.

When I pointed the code at "Currant", it worked *once*, but after that it
fails to even connect to the "Currant" AP (or the "Currant 5G" AP).