The relay coils controlled by the buttons were swapped
So, the right way to do this is to have "LED0" associated with "COIL2", and
"LED1" with "COIL1". So obviously it'd be a good idea to change the names
of these #defines, but they are consistent with the schematic's signal
names at the moment and we'll leave them like this through bringup.