Debugging the s24c08 eeprom chip reset/initialization implementation on PICO-KIT devboard
The s24c08 spec describes a very unusual set of contortions required on the
i2c bus in order to reset it (and apparently you have to reset the chip,
power-on alone isn't enough - but that's not 100% proved yet).

Still, it's worth a go to see if we can produce the reset conditions on the
i2c bus.

Moving to the PICO-KIT devboard for the moment because there I can attach
the Salea logic analyzer.