Multi-byte reads are supported (only up to the next eeprom page boundary)
So if you start at the address at the beginning of a page (either 0x000,
0x100, 0x200 or 0x300), you can read as many as 256 bytes.

But if you start at an intermediate location, the max you can read is 256
minus the offset of that location..

This isn't a fundamental limitation at all, just have to complicate the
code a little bit to support this (by re-establishing the read address each
time you cross into a new page's territory).

Will do it later.