Added "eeprom --blast --value=0xNN" to write a given value to all eeprom locations
Just to supplement the "eeprom --erase" command, that writes 0x00 to all of
the locations.

If MEASURE_EEPROM_WRITE_TIME is #defined, timing information will be
printed at the end of either a --blast or an --erase operation telling you
how long it took to write all 1024 bytes.

While I had this enabled, I was seeing times of 32 or 33 "ticks" (which I
believe means "milliseconds")