Now reporting time required to read entire eeprom
It's coming in at 30 milliseconds, consistent with my calculations:

\_\_\_\_ \_\_\_\_ \_\_\_\_ \_\_\_\_ \_\_\_\_ \_\_\_\_ \_\_\_\_ \_\_\_\_ \_\_\_\_

omar> eeprom -d
access_eeprom(): Reading all 1024 bytes of eeprom took 3 ticks (that's about 30 milliseconds, since 1 tick is 10 milliseconds)

\_\_\_\_ \_\_\_\_ \_\_\_\_ \_\_\_\_ \_\_\_\_ \_\_\_\_ \_\_\_\_ \_\_\_\_ \_\_\_\_

Note that writing the entire eeprom is much slower - about 40 times slower!
It takes 1.28 seconds to write to all 1024 bytes in eeprom, even using the
"fast" page write mode. This is because of the long value of
S24C08_WRITE_DELAY: 20 milliseconds. But there's not a lot we can do about
it; I've done experiments that indicate that 10 milliseconds is too
short (you see a high rate of write failures). So this seems to imply that
even if we tried the "acknowledge polling" feature that the most we could
hope to save would be around 0.6 seconds.