Start moving to pwm control of led brightness
First step: wired up all the pwm configuration, and running the "old"
console test commands for turning the leds on or off through the pwm apis.

The "duty_resolution" is set to LEDC_TIMER_13_BIT, or 8192 for now.
The frequency is 5000 Hz, meaning that each 200 usec timeslice is split up
into 8192 intervals that the PWM uses to dither the brightness.