Can now change the brightness of the leds by specifying the pwm duty cycle
The new console "pwm" command lets you exercise this.

In this implementation, the theoretical max duty cycle is 8192, but we
limit the max to 7936 (so as to leave a guaranteed slice of time during
each 200 usec pwm cycle where the led is off - 1/32 of the basic pwm
period, or about 6.4 usec).