Every 2 seconds a timer routine pauses the leds, takes and displays an als reading
I'm not sure if this is doing what it is supposed to, but checkpointing the
code for now since this is the way to go as far as overall structure goes.

Need to think about an ambient light sensor reading interface that avoids
contention on the i2c bus! Right now there's no way to guarantee that a
"foreground" activation of als reading code (using the console's "als"
command) won't collide with the background calls that are being made.