The sampling of the als adc takes place after a short interval controlled by secondary timer
The idea is that vTaskDelay() may not give us fine enough control over the
interval between the time the led pwm controller is paused/off, and the
time it is re-enabled to prevent us from seeing flickering.

We can make the chace of user-visible flickering even smaller by moving the
code that re-enables the led pwm controller into the timer_group0_isr()
routine.

ALSO: restored the 2 second delay between samplings.