The current pause-pwm-and-sneak-a-peek approach fails because pwm may pause while at a high value
This is a special build that slows down "primary" and "secondary"
omar_als_timer.c timers to 5 and 1 seconds, respectively.

At these slow speeds, the underlying cause of the confusing behavior I've
been seeing became clear: when ledc_timer_pause() is called, there is no
guarantee that the pwm will pause with the output 'low' -- in fact, the
brighter the pwm duty cycle setting, the greater the chance that you will
'ause with the output 'high'!

This is because the higher the duty cycle, the greater the proportion of
the time the pwm led output is high. And so, randomly pausing, the greater
the chance that it'll pause in the high state. And if you pause it for
"long time" (i.e., order of a second), to the human eye it appears as if
the brightness setting of the led has been maxed out -- because you've
essentially set it to 100% duty cycle, since the pwm isn't active anymore!

So it looks like calling ledc_timer_pause() isn't the way to go. Instead,
something more specific (like querying the current pwm duty cycle for LED1
and LED2 before setting the duty cycle to zero, and then restoring that
after you're done) would do the trick.