The "als" console command gives you (limited) ability to re-set the als timer periods
This feature is kind of ugly, but the idea is to let you experiment with
stretching out the secondary timer period (relative to the primary period)
so that you can see how much time you need to allow the phototransistor's
output to drop to the ambient level.

Previously, to run this kind of experiment, I had to change #defined
periods in omar_als_timer.h and re-build.. This change allows you to do it
on the fly.

The problem is, I can't get it to work on the "primary" timer interval. But
maybe that's not the important one... Might clean it up by removing that
option.