Replaced omar_als_timer.[ch] with the much simpler version I created yesterday
I took my own copy of the "timer_group" sample yesterday, and ripped the
code down to something very simple:using a single timer, I implemented a
two-period timer scheme that waits 1 second, and then waits 250 usecs. This
could be easily augmented to implement our ambient light sensor "cut and
peek" strategy. It's in this repo:

  /Users/vincecotter/Documents/work/currant/TASKS/TASK020_ESP32/gits/timer_group

This is the first step of porting this simplified code into omar. It
compiles and runs, but as soon as you enable the timer, it crashes from a
WDT exception. I think I know how to fix that, but checkpointing the code
as-is for now (it doesn't even read the als sensor yet :-)